feat: update SGLang performance collectors for v0.5.8 (#323)
* add perf data for SGLang0.5.8 on H100


* update wideep collectors for sgl0.5.8


* remove data files


---------
