+++ collector/sglang/collect_attn.py
@@ -36,7 +36,7 @@ from helper import benchmark_with_power, get_sm_version, log_perf
 
 DISABLE_BACKWARD = os.getenv("FLASH_ATTENTION_DISABLE_BACKWARD", "FALSE") == "TRUE"
 
-compatible_version = ["0.5.5.post3", "0.5.6.post2"]
+compatible_version = ["0.5.5.post3", "0.5.6.post2", "0.5.8"]
 
 
 class Timing(NamedTuple):
@@ -58,6 +58,7 @@ class MockModelConfig:
         self.full_attention_layer_ids = []
         self.is_multimodal = False
         self.hidden_size = num_attention_heads * head_dim
+        self.is_local_attention_model = False
 
         class MockHFConfig:
             def __init__(self):
